Change TA selection to dropdown menu
