+++ client/src/components/includes/ProfessorCalendarTable.tsx
@@ -4,6 +4,7 @@ import ProfessorCalendarRow from './ProfessorCalendarRow';
 class ProfessorCalendarTable extends React.Component {
 
     props: {
+        taList: string[],
         mondayList: string[][],
         tuesdayList: string[][],
         wednesdayList: string[][],
@@ -23,6 +24,7 @@ class ProfessorCalendarTable extends React.Component {
                         <th colSpan={tablewidth}>Monday</th>
                     </tr>
                     <ProfessorCalendarRow
+                        taList={this.props.taList}
                         time={this.props.mondayList[0]}
                         ta={this.props.mondayList[1]}
                         location={this.props.mondayList[2]}
@@ -32,6 +34,7 @@ class ProfessorCalendarTable extends React.Component {
                         <th colSpan={tablewidth}>Tuesday</th>
                     </tr>
                     <ProfessorCalendarRow
+                        taList={this.props.taList}
                         time={this.props.tuesdayList[0]}
                         ta={this.props.tuesdayList[1]}
                         location={this.props.tuesdayList[2]}
@@ -41,6 +44,7 @@ class ProfessorCalendarTable extends React.Component {
                         <th colSpan={tablewidth}>Wednesday</th>
                     </tr>
                     <ProfessorCalendarRow
+                        taList={this.props.taList}
                         time={this.props.wednesdayList[0]}
                         ta={this.props.wednesdayList[1]}
                         location={this.props.wednesdayList[2]}
@@ -50,6 +54,7 @@ class ProfessorCalendarTable extends React.Component {
                         <th colSpan={tablewidth}>Thursday</th>
                     </tr>
                     <ProfessorCalendarRow
+                        taList={this.props.taList}
                         time={this.props.thursdayList[0]}
                         ta={this.props.thursdayList[1]}
                         location={this.props.thursdayList[2]}
@@ -59,6 +64,7 @@ class ProfessorCalendarTable extends React.Component {
                         <th colSpan={tablewidth}>Friday</th>
                     </tr>
                     <ProfessorCalendarRow
+                        taList={this.props.taList}
                         time={this.props.fridayList[0]}
                         ta={this.props.fridayList[1]}
                         location={this.props.fridayList[2]}
@@ -68,6 +74,7 @@ class ProfessorCalendarTable extends React.Component {
                         <th colSpan={tablewidth}>Saturday</th>
                     </tr>
                     <ProfessorCalendarRow
+                        taList={this.props.taList}
                         time={this.props.saturdayList[0]}
                         ta={this.props.saturdayList[1]}
                         location={this.props.saturdayList[2]}
@@ -77,6 +84,7 @@ class ProfessorCalendarTable extends React.Component {
                         <th colSpan={tablewidth}>Sunday</th>
                     </tr>
                     <ProfessorCalendarRow
+                        taList={this.props.taList}
                         time={this.props.sundayList[0]}
                         ta={this.props.sundayList[1]}
                         location={this.props.sundayList[2]}
