+++ client/src/styles/index.less
@@ -25,6 +25,7 @@
 @import "./professor/ProfessorView";
 @import "./professor/ProfessorCalendarTable";
 @import "./professor/ProfessorCalendarRow";
+@import "./professor/ProfessorAddNewOH";
 
 #root {
   height:100%;
