+++ client/src/styles/professor/ProfessorAddNewOH.less
@@ -1,0 +1,115 @@
+.ProfessorAddNewOH {
+    .Add.false {
+        display: none;
+    }
+
+    .ExpandedAdd.false {
+        display: none;
+    }
+
+    .Add {
+        .NewOHButton {
+        background-color: white;
+        border: none;
+        padding: 15px;
+        margin: 30px 0px;
+        border-radius: 3px;
+        font-size: 14px;
+        box-shadow: 2px 2px 4px 0 rgba(158, 158, 158, 0.5);
+        float: left;
+        cursor: pointer;
+
+            .plus.icon {
+                margin-right: 10px;
+            }
+        }
+    }  
+
+    .ExpandedAdd {
+        margin: 30px 0px;
+        background-color: white;
+        width: 100%;
+        text-align: left;
+        box-shadow: 2px 2px 4px 0 rgba(158, 158, 158, 0.5);
+
+        .NewOHHeader {
+            border-bottom: 1px solid #d9d9d9;
+
+            .ExpandedNewOHButton {
+                background-color: white;
+                border: none;
+                padding: 15px;
+                border-radius: 3px;
+                font-size: 14px;
+                text-align: left;
+
+                .plus.icon {
+                    margin-right: 10px;
+                }
+            }            
+        }
+
+        .InfoInput {
+            margin-left: 40px;
+
+            .TA {
+                margin-top: 14px;
+    
+                .AddTAButton {
+                    color: dodgerblue;
+                    background-color: white;
+                    border: none;
+                }
+    
+                .long {
+                    width: 40%;
+                }
+            }
+    
+            .Location {
+                margin-top: 24px;
+    
+                .long {
+                    width: 40%;
+                }
+            }
+            
+            .Time {
+                margin-top: 24px;
+
+                .repeat {
+                    margin-left: 2%;
+                    color: black;
+                }
+            }
+        }
+    
+        .Buttons {
+            border-top: 1px solid #d9d9d9;
+            margin-top: 35px;
+            padding: 15px 0px;
+            background-color: white;
+            width: 100%;
+            text-align: right;
+
+            .Create, .Cancel {
+                padding: 10px 40px;
+                border-radius: 3px;
+                font-size: 14px;
+                margin: 0px 15px;
+            }
+
+            .Create {
+                background-color: dodgerblue;
+                color: white;
+                border: none;
+            }
+
+            .Cancel {
+                background-color: white;
+                border: 1px solid gray;
+            }
+
+        }
+    }
+}
