+++ client/src/styles/professor/ProfessorCalendarRow.less
@@ -15,9 +15,9 @@
 
 .Pair.true {
     border: 1px solid gray;
+    // padding-right: 10px; 
     .Preview {
         color: white;
-
         .Edit, .Delete {
             color: white;
         }
@@ -84,12 +84,8 @@
                 margin-top: 24px;
     
                 .repeat {
-                    width: 20px;
-                    height: 20px;
-                    margin: 0px 10px 0px 30px;
-                    border:2px solid black;
-                    background-color: white;
-                    border-radius: 3px;
+                    margin-left: 2%;
+                    color: black;
                 }
             }
         }
