+++ client/src/styles/professor/ProfessorView.less
@@ -1,31 +1,6 @@
 .ProfessorView {
     margin: 0px 100px;
 
-    .Add.false {
-        display: none;
-    }
-
-    .ExpandedAdd.false {
-        display: none;
-    }
-
-    .Add {
-        .NewOHButton {
-        background-color: white;
-        border: none;
-        padding: 15px;
-        margin: 30px 0px;
-        border-radius: 3px;
-        font-size: 14px;
-        box-shadow: 2px 2px 4px 0 rgba(158, 158, 158, 0.5);
-        float: left;   
-
-            .plus.icon {
-                margin-right: 10px;
-            }
-        }
-    }    
-
     input {
         border:none;
         background-color: #edeff2;
@@ -35,95 +10,28 @@
         width: 15%;
     }
 
-    .ExpandedAdd {
-        margin: 30px 0px;
-        background-color: white;
-        width: 100%;
-        text-align: left;
-        box-shadow: 2px 2px 4px 0 rgba(158, 158, 158, 0.5);
-
-        .NewOHHeader {
-            border-bottom: 1px solid #d9d9d9;
-
-            .ExpandedNewOHButton {
-                background-color: white;
-                border: none;
-                padding: 15px;
-                border-radius: 3px;
-                font-size: 14px;
-                text-align: left;
-
-                .plus.icon {
-                    margin-right: 10px;
-                }
-            }            
-        }
-
-        .InfoInput {
-            margin-left: 40px;
-
-            .TA {
-                margin-top: 14px;
-    
-                .AddTAButton {
-                    color: dodgerblue;
-                    background-color: white;
-                    border: none;
-                }
-    
-                .long {
-                    width: 40%;
-                }
-            }
-    
-            .Location {
-                margin-top: 24px;
-    
-                .long {
-                    width: 40%;
-                }
-            }
-            
-            .Time {
-                margin-top: 24px;
-
-                .repeat {
-                    width: 20px;
-                    height: 20px;
-                    margin: 0px 10px 0px 30px;
-                    border:2px solid black;
-                    background-color: white;
-                    border-radius: 3px;
-                }
-            }
-        }
-    
-        .Buttons {
-            border-top: 1px solid #d9d9d9;
-            margin-top: 35px;
-            padding: 15px 0px;
-            background-color: white;
-            width: 100%;
-            text-align: right;
-
-            .Create, .Cancel {
-                padding: 10px 40px;
-                border-radius: 3px;
-                font-size: 14px;
-                margin: 0px 15px;
-            }
-
-            .Create {
-                background-color: dodgerblue;
-                color: white;
-                border: none;
-            }
-
-            .Cancel {
-                background-color: white;
-                border: 1px solid gray;
-            }
-
-        }
+    Button {
+        cursor: pointer;
     }
+
+    // Dropdown CSS
+    .ui.selection.dropdown {
+        width: 40%;
+        margin: 0px 20px;
+        background-color: #edeff2;
+        border-radius: 0px;
+        border: none;
+        color: black;
+        padding: 5px 10px;
+        min-height: 25px;
+    }
+
+    .ui.selection.dropdown>.dropdown.icon{
+        padding: 4px 10px;
+    }
+    
+    .ui.dropdown:not(.button)>.default.text {
+        color: #898989;
+    }
+    // End Dropdown CSS
 }
